Rename folder for database migration
